removing kafka from auth microservice and unnecessary  consumers from profile and post. Added kafka to notifications microservice.
