+++ Microservices/Posts/src/main/java/cs4337/group_8/TemplateProject/controllers/KafkaProducerController.java
@@ -1,7 +1,7 @@
-package cs4337.group_8.AuthenticationMicroservice.controllers;
+package cs4337.group_8.TemplateProject.controllers;
 
 
-import cs4337.group_8.AuthenticationMicroservice.services.KafkaProducer;
+import cs4337.group_8.TemplateProject.services.KafkaProducer;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.bind.annotation.RequestMapping;
@@ -24,3 +24,4 @@ public class KafkaProducerController {
         return ResponseEntity.ok("Message sent to Kafka topic!");
     }
 }
+
