+++ Microservices/Posts/src/main/java/cs4337/group_8/TemplateProject/services/PostService.java
@@ -20,8 +20,13 @@ public class PostService {
     @Autowired
     private LikesRepository likesRepository;
 
+    @Autowired
+    private KafkaProducer kafkaProducer;
+
     public Post createPost(Post post) {
-        return postRepository.save(post);
+        Post savedPost = postRepository.save(post);
+        kafkaProducer.sendMessage("Post created with ID: " + savedPost.getId());
+        return savedPost;
     }
 
     public List<Post> getPostsByUser(Long userId) {
@@ -42,6 +47,7 @@ public class PostService {
 
         if (post.getUserId().equals(userId)) {
             postRepository.delete(post);
+            kafkaProducer.sendMessage("User ID " + userId + " Deleted Post ID " + postId);
         } else {
             throw new PostException("User ID " + userId + " is not authorized to delete post with ID " + postId);
         }
@@ -60,6 +66,7 @@ public class PostService {
                 likesRepository.save(like);
                 post.setLikesCount(post.getLikesCount() + 1); // Increment fast lookup counter
                 postRepository.save(post);
+                kafkaProducer.sendMessage("User ID " + userId + " liked Post ID " + postId);
             } else {
                 throw new PostException("User ID " + userId + " has already liked post with ID " + postId);
             }
@@ -69,6 +76,7 @@ public class PostService {
             likesRepository.save(newLike);
             post.setLikesCount(post.getLikesCount() + 1); // Increment fast lookup counter
             postRepository.save(post);
+            kafkaProducer.sendMessage("User ID " + userId + " liked Post ID " + postId);
         }
     }
 
@@ -97,6 +105,7 @@ public class PostService {
         }
 
         postRepository.save(post);
+        kafkaProducer.sendMessage("User ID " + userId + " unliked Post ID " + postId);
     }
 
     public long getLikeCount(Long postId) {
