+++ Microservices/Profile/src/main/resources/application.properties
@@ -19,7 +19,7 @@ logging.file.name=logs/app.log
 
 eureka.client.fetch-registry=true
 eureka.client.register-with-eureka=true
-eureka.client.service-url.defaultZone=http://localhost:8761/eureka/
+eureka.client.service-url.defaultZone=http://${SERVICE_REGISTRY_IP}:${PORT_SERVICE_REGISTRY}/eureka/
 eureka.instance.prefer-ip-address=true
 
 # Kafka Configuration
@@ -29,8 +29,3 @@ spring.kafka.bootstrap-servers=kafka:9092
 spring.kafka.producer.key-serializer=org.apache.kafka.common.serialization.StringSerializer
 spring.kafka.producer.value-serializer=org.apache.kafka.common.serialization.StringSerializer
 
-# Consumer Configurations
-spring.kafka.consumer.key-deserializer=org.apache.kafka.common.serialization.StringDeserializer
-spring.kafka.consumer.value-deserializer=org.apache.kafka.common.serialization.StringDeserializer
-spring.kafka.consumer.group-id=profile-group
-spring.kafka.consumer.auto-offset-reset=earliest
