+++ Microservices/notification-service/src/main/java/cs4337/group_8/NotificationService/controllers/KafkaConsumerController.java
@@ -1,7 +1,6 @@
-package cs4337.group_8.AuthenticationMicroservice.controllers;
+package cs4337.group_8.NotificationService.controllers;
 
-
-import cs4337.group_8.AuthenticationMicroservice.services.KafkaConsumer;
+import cs4337.group_8.NotificationService.services.KafkaConsumer;
 import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RestController;
